backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
alembic
pydantic
pydantic-settings
pydantic[email]
jinja2
weasyprint
redis
python-multipart
python-jose[cryptography]
passlib[bcrypt]
pytest
httpx
Pillow
qrcode[pil]
